ml/story_word_selector/requirements.txt
# Story Word Selector — ML Service
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0
supabase>=2.3.0
numpy>=1.26.0
httpx>=0.27.0
loguru>=0.7.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Shared ML infrastructure (cache + prediction logging + Celery)
redis>=5.0.0
celery[redis]>=5.4.0
